set up empty state in components
