+++ packages/core-components/src/lib/unsorted/viz/area/AreaChart.svelte
@@ -66,6 +66,9 @@
 
 	export let echartsOptions = undefined;
 	export let printEchartsConfig = false;
+
+	export let emptySet = undefined;
+	export let emptyMessage = undefined;
 </script>
 
 <Chart
@@ -101,6 +104,8 @@
 	{colorPalette}
 	{echartsOptions}
 	{printEchartsConfig}
+	{emptySet}
+	{emptyMessage}
 >
 	<Area
 		{line}
