+++ packages/core-components/src/lib/unsorted/viz/bubble/BubbleChart.svelte
@@ -55,6 +55,9 @@
 	export let colorPalette = undefined;
 	export let echartsOptions = undefined;
 	export let printEchartsConfig = false;
+
+	export let emptySet = undefined;
+	export let emptyMessage = undefined;
 </script>
 
 <Chart
@@ -92,6 +95,8 @@
 	{colorPalette}
 	{echartsOptions}
 	{printEchartsConfig}
+	{emptySet}
+	{emptyMessage}
 >
 	<Bubble {shape} {fillColor} {opacity} {outlineColor} {outlineWidth} {scaleTo} {useTooltip} />
 	<slot />
