+++ packages/core-components/src/lib/unsorted/viz/core/Chart.svelte
@@ -5,7 +5,22 @@
 <script>
 	import { QueryLoad } from '../../../atoms/query-load';
 	import Chart from './_Chart.svelte';
+	import EmptyChart from './EmptyChart.svelte';
+	import ErrorChart from './ErrorChart.svelte';
+
+	/** @type {import("@evidence-dev/query-store).QueryStore | unknown}*/
 	export let data;
+
+	const initialHash = typeof data === "object" && "__isQueryStore" in data ? data.hash : undefined;
+
+	let isInitial = data?.hash === initialHash;
+	$:  isInitial = data?.hash === initialHash;
+
+	/** @type {"pass" | "warn" | "error"}*/
+	export let emptySet = undefined;
+
+	/** @type {string}*/
+	export let emptyMessage = undefined;
 
 	// Remove any undefined props (e.g. w/o defaults) to prevent them from being passed
 	$: spreadProps = {
@@ -17,6 +32,8 @@
 
 <!-- Pass all the props through-->
 <QueryLoad {data} let:loaded>
+	<EmptyChart slot="empty" {emptyMessage} {emptySet} chartType={spreadProps.chartType} {isInitial}/>
+	<ErrorChart slot="error" chartType={spreadProps.chartType} error={loaded.error.message}/>
 	<Chart {...spreadProps} data={loaded?.__isQueryStore ? Array.from(loaded) : loaded} {queryID}>
 		<slot />
 	</Chart>
