+++ packages/core-components/src/lib/unsorted/viz/core/ErrorChart.svelte
@@ -17,10 +17,13 @@
 	class="grid grid-rows-auto box-content grid-cols-1 justify-center bg-red-50 text-grey-700 font-ui font-normal rounded border border-red-200 min-h-[150px] py-5 px-8 my-5 print:break-inside-avoid"
 >
 	<div class="m-auto w-full">
-		<div class="font-bold text-center text-lg">{chartType}</div>
+		<div 
+			class="font-bold text-center text-lg"
+		>
+			{chartType}
+		</div>
 		<div
 			class="text-center [word-wrap:break-work] text-xs"
-			class:w-[7.8em]={chartType.includes('Value')}
 		>
 			{error}
 			{#if dev && error === DevMissingCredentialsError}
