+++ packages/core-components/src/lib/unsorted/viz/core/_Value.svelte
@@ -6,7 +6,7 @@
 	} from '@evidence-dev/component-utilities/formatting';
 	import { convertColumnToDate } from '@evidence-dev/component-utilities/dateParsing';
 	import checkInputs from '@evidence-dev/component-utilities/checkInputs';
-
+	import ValueError from './ValueError.svelte';
 	import { strictBuild } from '@evidence-dev/component-utilities/chartContext';
 
 	// Passing in value from dataset:
@@ -86,6 +86,8 @@
 			}
 		} catch (e) {
 			error = e.message;
+			const setTextRed = '\x1b[31m%s\x1b[0m';
+			console.error(setTextRed, `Error in Value: ${error}`);
 			if (strictBuild) {
 				throw error;
 			}
@@ -102,15 +104,7 @@
 		{formatValue(selected_value, format_object)}
 	</span>
 {:else}
-	<span
-		class="group inline-flex items-center relative cursor-help cursor-helpfont-sans px-1 border border-red-200 py-[1px] bg-red-50 rounded"
-	>
-		<span class="inline font-sans font-medium text-xs text-red-600">error</span>
-		<span
-			class="hidden text-white font-sans group-hover:inline absolute -top-1 left-[105%] text-sm z-10 px-2 py-1 bg-gray-800/80 leading-relaxed min-w-[150px] max-w-[400px] rounded-md"
-			>{error}</span
-		>
-	</span>
+	<ValueError {error}/>
 {/if}
 
 <style>
