+++ packages/core-components/src/lib/unsorted/viz/histogram/Histogram.svelte
@@ -34,6 +34,9 @@
 	export let colorPalette = undefined;
 	export let echartsOptions = undefined;
 	export let printEchartsConfig = false;
+
+	export let emptySet = undefined;
+	export let emptyMessage = undefined;
 </script>
 
 <Chart
@@ -61,6 +64,8 @@
 	{colorPalette}
 	{echartsOptions}
 	{printEchartsConfig}
+	{emptySet}
+	{emptyMessage}
 >
 	<Hist {fillColor} {fillOpacity} />
 	<slot />
