+++ packages/core-components/src/lib/unsorted/viz/index.js
@@ -3,6 +3,7 @@ export { default as Area } from './area/Area.svelte';
 export { default as BarChart } from './bar/BarChart.svelte';
 export { default as Bar } from './bar/Bar.svelte';
 export { default as BigValue } from './core/BigValue.svelte';
+export { default as BigValueError } from './core/BigValueError.svelte';
 export { default as BoxPlot } from './box/BoxPlot.svelte';
 export { default as BubbleChart } from './bubble/BubbleChart.svelte';
 export { default as Bubble } from './bubble/Bubble.svelte';
@@ -13,6 +14,7 @@ export { default as DataTable } from './table/DataTable.svelte';
 export { default as EChartsCopyTarget } from './core/EChartsCopyTarget.svelte';
 export { default as EChartsMap } from './map/EChartsMap.svelte';
 export { default as ECharts } from './core/ECharts.svelte';
+export { default as EmptyChart } from './core/EmptyChart.svelte';
 export { default as ErrorChart } from './core/ErrorChart.svelte';
 export { default as FunnelChart } from './funnel/FunnelChart.svelte';
 export { default as Heatmap } from './heatmap/Heatmap.svelte';
@@ -29,4 +31,5 @@ export { default as ReferenceLine } from './references/ReferenceLine.svelte';
 export { default as ReferenceArea } from './references/ReferenceArea.svelte';
 export { default as USMap } from './map/USMap.svelte';
 export { default as Value } from './core/Value.svelte';
+export { default as ValueError } from './core/ValueError.svelte';
 export { default as VennDiagram } from './venn/VennDiagram.svelte';
