+++ packages/core-components/src/lib/unsorted/viz/line/LineChart.svelte
@@ -83,6 +83,9 @@
 
 	export let echartsOptions = undefined;
 	export let printEchartsConfig = false;
+
+	export let emptySet = undefined;
+	export let emptyMessage = undefined;
 </script>
 
 <Chart
@@ -127,6 +130,8 @@
 	{colorPalette}
 	{echartsOptions}
 	{printEchartsConfig}
+	{emptySet}
+	{emptyMessage}
 >
 	<Line
 		{lineColor}
