+++ packages/core-components/src/lib/unsorted/viz/map/USMap.svelte
@@ -5,7 +5,23 @@
 <script>
 	import { QueryLoad } from '../../../atoms/query-load';
 	import USMap from './_USMap.svelte';
+	import EmptyChart from '../core/EmptyChart.svelte';
+	import ErrorChart from '../core/ErrorChart.svelte';
+
 	export let data;
+
+	const initialHash = typeof data === "object" && "__isQueryStore" in data ? data.hash : undefined;
+
+	let isInitial = data?.hash === initialHash;
+	$:  isInitial = data?.hash === initialHash;
+
+	/** @type {"pass" | "warn" | "error"}*/
+	export let emptySet = undefined;
+
+	/** @type {string}*/
+	export let emptyMessage = undefined;
+
+	let chartType = "US Map"
 
 	// Remove any undefined props (e.g. w/o defaults) to prevent them from being passed
 	$: spreadProps = Object.fromEntries(Object.entries($$props).filter(([, v]) => v !== undefined));
@@ -15,6 +31,8 @@
 
 <!-- Pass all the props through-->
 <QueryLoad {data} let:loaded>
+	<EmptyChart slot="empty" {emptyMessage} {emptySet} {chartType} {isInitial}/>
+	<ErrorChart slot="error" {chartType} error={loaded.error.message}/>
 	<USMap {...spreadProps} data={loaded?.__isQueryStore ? Array.from(loaded) : loaded} {queryID}>
 		<slot />
 	</USMap>
