+++ packages/core-components/src/lib/unsorted/viz/scatter/ScatterPlot.svelte
@@ -54,6 +54,9 @@
 	export let colorPalette = undefined;
 	export let echartsOptions = undefined;
 	export let printEchartsConfig = false;
+
+	export let emptySet = undefined;
+	export let emptyMessage = undefined;
 </script>
 
 <Chart
@@ -90,6 +93,8 @@
 	{colorPalette}
 	{echartsOptions}
 	{printEchartsConfig}
+	{emptySet}
+	{emptyMessage}
 >
 	<Scatter {shape} {fillColor} {opacity} {outlineColor} {outlineWidth} {pointSize} {useTooltip} />
 	<slot />
